two funcs to utils
